Split tests Fix incorrect mocking
